create util.mantid module re #306
