+++ mslice/app/__init__.py
@@ -2,11 +2,8 @@
 and entry points.
 """
 
-from mantid.api import AlgorithmFactory
 from mslice.util.qt.QtWidgets import QApplication
-from mantid.simpleapi import _translate
-from mslice.models.projection.powder.make_projection import MakeProjection
-
+from mslice.util.mantid import initialize_mantid
 
 # Module-level reference to keep main window alive after show_gui has returned
 MAIN_WINDOW = None
@@ -23,8 +20,7 @@ def show_gui():
     If this is the first call then an instance of the Windows is cached to ensure it
     survives for the duration of the application
     """
-    AlgorithmFactory.subscribe(MakeProjection)
-    _translate()
+    initialize_mantid()
     global MAIN_WINDOW
     if MAIN_WINDOW is None:
         from mslice.app.mainwindow import MainWindow
